refactor(View): move package
