+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handleradapter/HandlerAdapter.java
@@ -1,6 +1,6 @@
 package nextstep.mvc.tobe.handleradapter;
 
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.view.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
